Front End/requirements.txt
absl-py==1.3.0
accelerate==1.2.1
aiohttp==3.9.5
aiosignal==1.3.1
albucore==0.0.15
albumentations==1.4.15
altair==5.5.0
annotated-types==0.7.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.2
asttokens==2.4.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bidict==0.23.1
bleach==6.0.0
blinker==1.9.0
cachetools==5.2.0
cassandra-driver==3.29.2
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==3.0.0
cmdstanpy==1.2.5
colorama==0.4.5
comm==0.1.4
comtypes==1.1.14
contourpy==1.0.5
convertdate==2.4.0
cors==1.0.1
cycler==0.11.0
dataclasses-json==0.6.7
dearpygui==1.11.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
Django==4.1.2
dnspython==2.7.0
docker-pycreds==0.4.0
ephem==4.1.6
eval_type_backport==0.2.0
exceptiongroup==1.1.3
executing==1.2.0
face-recognition-models==0.3.0
Farama-Notifications==0.0.4
fastjsonschema==2.18.0
filelock==3.13.1
Flask==2.2.2
Flask-Cors==5.0.0
Flask-SocketIO==5.5.0
flatbuffers==24.3.25
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
future==0.18.3
gast==0.4.0
geographiclib==2.0
geomet==0.2.1.post1
geopy==2.4.1
gevent==23.9.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.66.0
googletrans==4.0.0rc1
graphviz==0.20.3
greenlet==3.0.1
grpcio==1.50.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
h11==0.9.0
h2==3.2.0
h5py==3.11.0
holidays==0.63
hpack==3.0.0
hstspreload==2024.12.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
huggingface-hub==0.27.0
hyperframe==5.2.0
idna==2.10
image==1.5.33
imageio==2.35.1
importlib_resources==6.4.5
imutils==0.5.4
instagram-private-api==1.6.0.0
intel-openmp==2021.4.0
ipykernel==6.25.2
ipython==8.15.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.4.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
keras==2.10.0
Keras-Preprocessing==1.1.2
keras-resnet==0.2.0
kiwisolver==1.4.4
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.10
langchain-openai==0.1.11
langchain-text-splitters==0.2.2
langsmith==0.1.82
lazy_loader==0.4
libclang==14.0.6
loguru==0.7.2
LunarCalendar==0.0.9
Markdown==3.4.1
markdown-it-py==3.0.0
markov-clustering==0.0.6.dev0
MarkupSafe==2.1.1
marshmallow==3.21.3
matplotlib==3.9.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
mkl==2021.4.0
ml-dtypes==0.4.1
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
namex==0.0.8
narwhals==1.19.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.3
nltk==3.9.1
notebook==7.0.3
notebook_shim==0.2.3
numpy==1.26.4
oauthlib==3.2.2
openai==1.35.7
opencv-python==4.8.1.78
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.5
overrides==7.4.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.3
patsy==1.0.1
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==3.10.0
prettytable==0.7.2
prometheus-client==0.17.1
prompt-toolkit==3.0.39
prophet==1.1.4
proto-plus==1.25.0
protobuf==3.20.0
psutil==5.9.5
psycopg2==2.9.10
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==18.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.54
pycocotools==2.0.8
pycparser==2.21
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
pygame==2.1.3
PyGetWindow==0.0.9
Pygments==2.16.1
PyMeeus==0.5.12
pymongo==4.10.1
PyMsgBox==1.0.9
PyMySQL==1.1.1
pyparsing==3.0.9
pyperclip==1.8.2
pypiwin32==223
pyreadline==2.1
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.11.1
python-json-logger==2.0.7
python-socketio==5.12.0
pyttsx3==2.90
pytweening==1.0.7
pytz==2022.7.1
pywin32==304
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
regex==2024.5.15
requests==2.32.3
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.10.3
rsa==4.9
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.2
sentencepiece==0.2.0
sentry-sdk==1.9.0
setproctitle==1.3.3
simple-websocket==1.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
speedtest==0.0.1
speedtest-cli==2.1.3
SQLAlchemy==2.0.31
sqlparse==0.4.3
stable_baselines3==2.3.2
stack-data==0.6.2
stanio==0.5.1
statsmodels==0.14.4
streamlit==1.41.1
sumolib==1.20.0
sympy==1.12.1
tbb==2021.13.0
tenacity==8.4.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.18.0
tensorflow-cpu==2.16.1
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.1
tensorflow-io-gcs-filesystem==0.27.0
tensorflow_intel==2.18.0
termcolor==2.0.1
terminado==0.17.1
tf-estimator-nightly==2.8.0.dev2021122109
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.8.30
tiktoken==0.7.0
tinycss2==1.2.1
tkcalendar==1.6.1
tldextract==5.1.1
tokenizers==0.21.0
toml==0.10.2
tomli==2.0.1
torch==2.4.1
torchvision==0.19.1
tornado==6.3.3
tqdm==4.66.4
traci==1.20.0
traitlets==5.9.0
transformers==4.47.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.0.145
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.25.11
VehicleDetectionTracker==0.0.32
waitress==3.0.2
wandb==0.18.1
watchdog==6.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
Werkzeug==2.2.2
win32-setctime==1.1.0
wrapt==1.14.1
wsproto==1.2.0
xgboost==2.1.3
yarl==1.9.4
yt-dlp==2024.12.13
zope.event==5.0
zope.interface==6.1
